// app/frontend/src/app/models/rezervacija.ts
export class Rezervacija {
  idRez: number = 0;
  datum: string= "";
  vreme: string = '';
  brOsoba: number = 0;
  opis: string = '';
  status: string = '';
  idR: number = 0;
  porukaRadnika: string = '';
  idSto: number = 0;
  nazivR: string = '';
  adresaR: string = '';
  radnik_kor_ime: string = "";

}